rag_requirements.txt
langchain 
langchain-openai 
langchain-community 
langchain-core 
faiss-cpu 
python-dotenv
scikit-learn 
nltk
streamlit